Fix but in parsing of hyperlinks.
svn path=/trunk/koffice/; revision=1079596

+++ filters/kpresenter/powerpoint/mso/simpleParser.h
@@ -1858,13 +1858,13 @@ public:
 class TargetAtom : public StreamOffset {
 public:
     OfficeArtRecordHeader rh;
-    QVector<quint32> target;
+    QVector<quint16> target;
     TargetAtom(void* /*dummy*/ = 0) {}
 };
 class LocationAtom : public StreamOffset {
 public:
     OfficeArtRecordHeader rh;
-    QVector<quint32> location;
+    QVector<quint16> location;
     LocationAtom(void* /*dummy*/ = 0) {}
 };
 class ExMCIMovieContainer : public StreamOffset {
